+++ src/base.ts
@@ -468,26 +468,10 @@ export class JSONApiResponse<T> {
 	}
 }
 
-export class VoidApiResponse {
-	constructor(public raw: Response) {}
+// export class TextApiResponse {
+// 	constructor(public raw: Response) {}
 
-	async value(): Promise<void> {
-		return undefined;
-	}
-}
-
-export class BlobApiResponse {
-	constructor(public raw: Response) {}
-
-	async value(): Promise<Blob> {
-		return await this.raw.blob();
-	}
-}
-
-export class TextApiResponse {
-	constructor(public raw: Response) {}
-
-	async value(): Promise<string> {
-		return await this.raw.text();
-	}
-}
+// 	async value(): Promise<string> {
+// 		return await this.raw.text();
+// 	}
+// }
